+++ tests/ValidateTest.php
@@ -1,234 +1,304 @@
 <?php
 
 /**
- * @package Pulsar
  * @author Jared King <j@jaredtking.com>
- * @link http://jaredtking.com
+ *
+ * @see http://jaredtking.com
+ *
  * @copyright 2015 Jared King
  * @license MIT
  */
-
 use Infuse\Utility;
-use Pulsar\Validate;
+use Pulsar\Validator;
 
 class ValidateTest extends PHPUnit_Framework_TestCase
 {
+    public function testGetRules()
+    {
+        $validator = new Validator('alpha');
+        $this->assertEquals('alpha', $validator->getRules());
+    }
+
     public function testAlpha()
     {
+        $validator = new Validator('alpha');
         $s = 'abc';
-        $this->assertTrue(Validate::is($s, 'alpha'));
+        $this->assertTrue($validator->validate($s));
         $s = ')S*F#$)S*';
-        $this->assertFalse(Validate::is($s, 'alpha'));
+        $this->assertFalse($validator->validate($s));
+
+        $validator = new Validator('alpha:5');
         $s = 'abcde';
-        $this->assertTrue(Validate::is($s, 'alpha:5'));
+        $this->assertTrue($validator->validate($s));
         $s = 'abc';
-        $this->assertFalse(Validate::is($s, 'alpha:5'));
+        $this->assertFalse($validator->validate($s));
     }
 
     public function testAlphaNumeric()
     {
+        $validator = new Validator('alpha_numeric');
         $s = 'abc1234';
-        $this->assertTrue(Validate::is($s, 'alpha_numeric'));
+        $this->assertTrue($validator->validate($s));
         $s = ')S*F#$)S*';
-        $this->assertFalse(Validate::is($s, 'alpha_numeric'));
+        $this->assertFalse($validator->validate($s));
+
+        $validator = new Validator('alpha_numeric:5');
         $s = 'a2cde';
-        $this->assertTrue(Validate::is($s, 'alpha_numeric:5'));
+        $this->assertTrue($validator->validate($s));
         $s = 'a2c';
-        $this->assertFalse(Validate::is($s, 'alpha_numeric:5'));
+        $this->assertFalse($validator->validate($s));
     }
 
     public function testAlphaDash()
     {
+        $validator = new Validator('alpha_dash');
         $s = 'abc-1234';
-        $this->assertTrue(Validate::is($s, 'alpha_dash'));
+        $this->assertTrue($validator->validate($s));
         $s = ')S*F#$)S*';
-        $this->assertFalse(Validate::is($s, 'alpha_dash'));
+        $this->assertFalse($validator->validate($s));
+
+        $validator = new Validator('alpha_dash:5');
         $s = 'r2-d2';
-        $this->assertTrue(Validate::is($s, 'alpha_dash:5'));
-        $this->assertFalse(Validate::is($s, 'alpha_dash:7'));
+        $this->assertTrue($validator->validate($s));
+
+        $validator = new Validator('alpha_dash:7');
+        $this->assertFalse($validator->validate($s));
     }
 
     public function testBoolean()
     {
+        $validator = new Validator('boolean');
+
         $s = '1';
-        $this->assertTrue(Validate::is($s, 'boolean'));
+        $this->assertTrue($validator->validate($s));
         $this->assertTrue($s);
         $s = '0';
-        $this->assertTrue(Validate::is($s, 'boolean'));
+        $this->assertTrue($validator->validate($s));
         $this->assertFalse($s);
     }
 
     public function testEmail()
     {
+        $validator = new Validator('email');
         $s = 'test@example.com';
-        $this->assertTrue(Validate::is($s, 'email'));
+        $this->assertTrue($validator->validate($s));
         $s = 'test';
-        $this->assertFalse(Validate::is($s, 'email'));
+        $this->assertFalse($validator->validate($s));
     }
 
     public function testEnum()
     {
+        $validator = new Validator('enum:red,orange,yellow,green,blue,violet');
         $s = 'blue';
-        $this->assertTrue(Validate::is($s, 'enum:red,orange,yellow,green,blue,violet'));
+        $this->assertTrue($validator->validate($s));
+
+        $validator = new Validator('enum:Austin,Dallas,OKC,Tulsa');
         $s = 'Paris';
-        $this->assertFalse(Validate::is($s, 'enum:Austin,Dallas,OKC,Tulsa'));
+        $this->assertFalse($validator->validate($s));
     }
 
     public function testDate()
     {
         date_default_timezone_set('UTC');
+
+        $validator = new Validator('date');
         $s = 'today';
-        $this->assertTrue(Validate::is($s, 'date'));
+        $this->assertTrue($validator->validate($s));
         $s = '09/17/2013';
-        $this->assertTrue(Validate::is($s, 'date'));
+        $this->assertTrue($validator->validate($s));
         $s = 'doesnotwork';
-        $this->assertFalse(Validate::is($s, 'date'));
+        $this->assertFalse($validator->validate($s));
     }
 
     public function testIp()
     {
+        $validator = new Validator('ip');
         $s = '127.0.0.1';
-        $this->assertTrue(Validate::is($s, 'ip'));
+        $this->assertTrue($validator->validate($s));
         $s = 'doesnotwork';
-        $this->assertFalse(Validate::is($s, 'ip'));
+        $this->assertFalse($validator->validate($s));
     }
 
     public function testMatching()
     {
+        $validator = new Validator('matching');
         $match = 'notarray';
-        $this->assertTrue(Validate::is($match, 'matching'));
+        $this->assertTrue($validator->validate($match));
 
         $match = ['test', 'test'];
-        $this->assertTrue(Validate::is($match, 'matching'));
+        $this->assertTrue($validator->validate($match));
         $this->assertEquals('test', $match);
 
         $match = ['test', 'test', 'test', 'test'];
-        $this->assertTrue(Validate::is($match, 'matching'));
+        $this->assertTrue($validator->validate($match));
         $this->assertEquals('test', $match);
 
         $notmatching = ['test', 'nope'];
-        $this->assertFalse(Validate::is($notmatching, 'matching'));
+        $this->assertFalse($validator->validate($notmatching));
         $this->assertEquals(['test', 'nope'], $notmatching);
     }
 
     public function testNumeric()
     {
+        $validator = new Validator('numeric');
         $s = 12345.22;
-        $this->assertTrue(Validate::is($s, 'numeric'));
+        $this->assertTrue($validator->validate($s));
         $s = '1234';
-        $this->assertTrue(Validate::is($s, 'numeric'));
+        $this->assertTrue($validator->validate($s));
         $s = 'notanumber';
-        $this->assertFalse(Validate::is($s, 'numeric'));
+        $this->assertFalse($validator->validate($s));
+
+        $validator = new Validator('numeric:double');
         $s = 12345.22;
-        $this->assertTrue(Validate::is($s, 'numeric:double'));
+        $this->assertTrue($validator->validate($s));
+
+        $validator = new Validator('numeric:int');
         $s = 12345.22;
-        $this->assertFalse(Validate::is($s, 'numeric:int'));
+        $this->assertFalse($validator->validate($s));
     }
 
     public function testPassword()
     {
         $salt = 'saltvalue';
-        Validate::configure(['salt' => $salt]);
+        Validator::configure(['salt' => $salt]);
 
+        $validator = new Validator('password:8');
         $password = 'testpassword';
-        $this->assertTrue(Validate::is($password, 'password:8'));
+        $this->assertTrue($validator->validate($password));
         $this->assertEquals(Utility::encryptPassword('testpassword', $salt), $password);
 
         $invalid = '...';
-        $this->assertFalse(Validate::is($invalid, 'password:8'));
+        $this->assertFalse($validator->validate($invalid));
     }
 
     public function testRange()
     {
         $s = -1;
-        $this->assertTrue(Validate::is($s, 'range'));
-        $this->assertTrue(Validate::is($s, 'range:-1'));
-        $this->assertTrue(Validate::is($s, 'range:-1:100'));
+        $validator = new Validator('range');
+        $this->assertTrue($validator->validate($s));
+
+        $validator = new Validator('range:-1');
+        $this->assertTrue($validator->validate($s));
+
+        $validator = new Validator('range:-1:100');
+        $this->assertTrue($validator->validate($s));
 
         $s = 100;
-        $this->assertFalse(Validate::is($s, 'range:101'));
-        $this->assertFalse(Validate::is($s, 'range:0:99'));
+        $validator = new Validator('range:101');
+        $this->assertFalse($validator->validate($s));
+
+        $validator = new Validator('range:0:99');
+        $this->assertFalse($validator->validate($s));
     }
 
     public function testRequired()
     {
+        $validator = new Validator('required');
         $s = 'ok';
-        $this->assertTrue(Validate::is($s, 'required'));
+        $this->assertTrue($validator->validate($s));
         $s = '';
-        $this->assertFalse(Validate::is($s, 'required'));
+        $this->assertFalse($validator->validate($s));
     }
 
     public function testString()
     {
         $s = 'thisisok';
-        $this->assertTrue(Validate::is($s, 'string'));
-        $this->assertTrue(Validate::is($s, 'string:5'));
-        $this->assertTrue(Validate::is($s, 'string:1:8'));
-        $this->assertTrue(Validate::is($s, 'string:0:9'));
-        $this->assertFalse(Validate::is($s, 'string:9'));
-        $this->assertFalse(Validate::is($s, 'string:1:7'));
+        $validator = new Validator('string');
+        $this->assertTrue($validator->validate($s));
+
+        $validator = new Validator('string:5');
+        $this->assertTrue($validator->validate($s));
+        $validator = new Validator('string:1:8');
+        $this->assertTrue($validator->validate($s));
+        $validator = new Validator('string:0:9');
+        $this->assertTrue($validator->validate($s));
+        $validator = new Validator('string:9');
+        $this->assertFalse($validator->validate($s));
+        $validator = new Validator('string:1:7');
+        $this->assertFalse($validator->validate($s));
 
         $s = new stdClass();
-        $this->assertFalse(Validate::is($s, 'string'));
+        $validator = new Validator('string');
+        $this->assertFalse($validator->validate($s));
     }
 
     public function testTimeZone()
     {
+        $validator = new Validator('time_zone');
         $s = 'America/Chicago';
-        $this->assertTrue(Validate::is($s, 'time_zone'));
+        $this->assertTrue($validator->validate($s));
+
         $s = 'anywhere';
-        $this->assertFalse(Validate::is($s, 'time_zone'));
+        $this->assertFalse($validator->validate($s));
     }
 
     public function testTimestamp()
     {
+        $validator = new Validator('timestamp');
         $s = $t = time();
-        $this->assertTrue(Validate::is($s, 'timestamp'));
+        $this->assertTrue($validator->validate($s));
         $this->assertEquals($t, $s);
 
         $s = 'today';
-        $this->assertTrue(Validate::is($s, 'timestamp'));
+        $this->assertTrue($validator->validate($s));
         $this->assertEquals(strtotime('today'), $s);
     }
 
     public function testDbTimestamp()
     {
+        $validator = new Validator('db_timestamp');
+
         $s = mktime(23, 34, 20, 4, 18, 2012);
-        $this->assertTrue(Validate::is($s, 'db_timestamp'));
+        $this->assertTrue($validator->validate($s));
         $this->assertEquals('2012-04-18 23:34:20', $s);
 
         $s = 'test';
-        $this->assertFalse(Validate::is($s, 'db_timestamp'));
+        $this->assertFalse($validator->validate($s));
     }
 
     public function testUrl()
     {
+        $validator = new Validator('url');
+
         $s = 'http://example.com';
-        $this->assertTrue(Validate::is($s, 'url'));
+        $this->assertTrue($validator->validate($s));
         $s = 'notaurl';
-        $this->assertFalse(Validate::is($s, 'url'));
+        $this->assertFalse($validator->validate($s));
     }
 
-    public function testMultipleRequirements()
+    public function testMultipleRules()
     {
+        $validator = new Validator('matching|string:2');
+
         $t = ['test', 'test'];
-        $this->assertTrue(Validate::is($t, 'matching|string:2'));
+        $this->assertTrue($validator->validate($t));
         $this->assertEquals('test', $t);
     }
 
-    public function testKeyValueRequirements()
+    public function testKeyValueRules()
     {
         $test = [
             'test' => ['test', 'test'],
             'test2' => 'alphanumer1c',
         ];
 
-        $requirements = [
+        $rules = [
             'test' => 'matching|string:2',
             'test2' => 'alpha_numeric',
         ];
 
-        $this->assertTrue(Validate::is($test, $requirements));
+        $validator = new Validator($rules);
+
+        $this->assertTrue($validator->validate($test));
         $this->assertEquals('test', $test['test']);
+    }
+
+    public function testGetFailingRule()
+    {
+        $validator = new Validator('matching|string:5');
+
+        $t = ['test', 'test'];
+        $this->assertFalse($validator->validate($t));
+        $this->assertEquals('string', $validator->getFailingRule());
     }
 }